Implement photo service

+++ RoadDefectsService.Core.Application/Services/PhotoService.cs
@@ -2,6 +2,7 @@
 using RoadDefectsService.Core.Application.Interfaces.Repositories;
 using RoadDefectsService.Core.Application.Interfaces.Services;
 using RoadDefectsService.Core.Application.Models;
+using RoadDefectsService.Core.Domain.Enums;
 using RoadDefectsService.Core.Domain.Models;
 
 namespace RoadDefectsService.Core.Application.Services
@@ -10,18 +11,25 @@ namespace RoadDefectsService.Core.Application.Services
     {
         private readonly IFileService _fileService;
         private readonly IPhotoRepository _photoRepository;
+        private readonly IFixationWorkRepository _fixationWorkRepository;
+        private readonly IFixationDefectRepository _fixationDefectRepository;
 
-        public PhotoService(IFileService fileService, IPhotoRepository photoRepository)
+        public PhotoService(
+            IFileService fileService, IPhotoRepository photoRepository,
+            IFixationWorkRepository fixationWorkRepository, IFixationDefectRepository fixationDefectRepository)
         {
             _fileService = fileService;
             _photoRepository = photoRepository;
+            _fixationWorkRepository = fixationWorkRepository;
+            _fixationDefectRepository = fixationDefectRepository;
         }
 
-        public async Task<ExecutionResult<PhotoUploadResponseDTO>> AddPhotoAsync(PhotoDTO addPhoto)
+        public async Task<ExecutionResult<PhotoUploadResponseDTO>> AddPhotoAsync(PhotoDTO addPhoto, Guid ownerId)
         {
             Photo photo = new()
             {
                 Id = Guid.NewGuid(),
+                OwnerId = ownerId,
                 Name = addPhoto.Name,
                 Type = addPhoto.Type,
             };
@@ -37,12 +45,23 @@ namespace RoadDefectsService.Core.Application.Services
             return new PhotoUploadResponseDTO() { UploadedPhotoId = photo.Id };
         }
 
-        public async Task<ExecutionResult> DeletePhotoAsync(Guid photoId)
+        public async Task<ExecutionResult> DeletePhotoAsync(Guid photoId, Guid? ownerId)
         {
             Photo? photo = await _photoRepository.GetByIdAsync(photoId);
             if (photo is null)
             {
                 return new(StatusCodeExecutionResult.NotFound, "PhotoNotFound", $"Photo with id {photoId} not found!");
+            }
+
+            if (ownerId.HasValue && photo.OwnerId != ownerId.Value)
+            {
+                return new(StatusCodeExecutionResult.Forbid, "DeletePhotoFail", $"You can't delete a photo because you don't own it!");
+            }
+
+            ExecutionResult checkResult = await CheckPhotoReferentsAsync(photo);
+            if (checkResult.IsNotSuccess)
+            {
+                return checkResult;
             }
 
             bool deleteResult = _fileService.DeleteFile(photo.PathName);
@@ -56,12 +75,58 @@ namespace RoadDefectsService.Core.Application.Services
             return ExecutionResult.SucceededResult;
         }
 
-        public async Task<ExecutionResult<PhotoDTO>> GetPhotoAsync(Guid photoId)
+        private Task<ExecutionResult> CheckPhotoReferentsAsync(Photo photo)
+        {
+            if (photo.FixationDefectId is not null)
+            {
+                return CheckPhotoReferentsAsync(
+                    photo.FixationDefectId.Value, 
+                    _fixationDefectRepository.GetByIdWithTaskAsync, 
+                    fixation => fixation.Task!.TaskStatus);
+            }
+            else if (photo.FixationWorkId is not null)
+            {
+                return CheckPhotoReferentsAsync(
+                    photo.FixationWorkId.Value, 
+                    _fixationWorkRepository.GetByIdWithTaskAsync,
+                    fixation => fixation.TaskFixationWork!.TaskStatus);
+            }
+
+            return Task.FromResult(ExecutionResult.SucceededResult);
+        }
+
+        private async Task<ExecutionResult> CheckPhotoReferentsAsync<TFixation>(Guid fixationId, Func<Guid, Task<TFixation?>> getByIdAsync, Func<TFixation, StatusTask> getStatus)
+        {
+            TFixation? fixation = await getByIdAsync(fixationId);
+            if (fixation is null)
+            {
+                return new(StatusCodeExecutionResult.NotFound, "FixationNotFound", $"Fixation with id {fixationId} not found!");
+            }
+
+            StatusTask status = getStatus(fixation);
+            if (status == StatusTask.Completed)
+            {
+                return new(StatusCodeExecutionResult.BadRequest, "TaskCompleted", "You cannot modify a completed task.");
+            }
+            else if (status == StatusTask.Created)
+            {
+                return new(StatusCodeExecutionResult.BadRequest, "TaskNotProcessing", "You cannot change a task that has not been started.");
+            }
+
+            return ExecutionResult.SucceededResult;
+        }
+
+        public async Task<ExecutionResult<PhotoDTO>> GetPhotoAsync(Guid photoId, Guid? ownerId)
         {
             Photo? photo = await _photoRepository.GetByIdAsync(photoId);
             if (photo is null)
             {
                 return new(StatusCodeExecutionResult.NotFound, "PhotoNotFound", $"Photo with id {photoId} not found!");
+            }
+
+            if (ownerId.HasValue && photo.OwnerId != ownerId.Value)
+            {
+                return new(StatusCodeExecutionResult.Forbid, "DeletePhotoFail", $"You can't delete a photo because you don't own it!");
             }
 
             byte[]? photoFile = await _fileService.GetFileAsync(photo.PathName);
